tests/unit/steam_reader: use size_t for _remaining_size
because not all platform use the same type for size_t and uint64_t, to be
on the safe side, and more error-prone, we'd better use the same type.

this change helps to address the following FTBFS:

tests/unit/stream_reader_test.cc:58:28: error: no matching function for call to 'min'
        _remaining_size -= std::min(_remaining_size, n);
                           ^~~~~~~~

+++ tests/unit/stream_reader_test.cc
@@ -39,7 +39,7 @@ using namespace util;
 class test_source_impl : public data_source_impl {
     short _current_letter = 0; // a-z corresponds to 0-25
     size_t _buffer_size;
-    size_t _remaining_size;
+    uint64_t _remaining_size;
 public:
     test_source_impl(size_t buffer_size, size_t total_size)
         : _buffer_size(buffer_size), _remaining_size(total_size) {
